requirements.txt
SQLAlchemy~=1.4.20
fastapi~=0.66.0
pydantic~=1.8.2
passlib~=1.7.4
uvicorn~=0.14.0
requests~=2.25.1
python-jose~=3.3.0
psycopg2_binary~=2.9.1
email-validator>=1.0.3
python-multipart~=0.0.5
bcrypt~=3.2.0
aiofiles==0.7.0
